src/requirements.txt
antlr4-python3-runtime==4.13.2
antlr4-tools==0.2.1
graphviz==0.20.3
install-jdk==1.1.0
llvmlite==0.43.0
